articles/requirements.txt
Django==2.1.4
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
python-dotenv==0.10.0
pytz==2018.7
